Moved entities to app.core.models
