+++ application/backend/app/core/models/shape.py
