--- application/backend/app/entities/__init__.py
@@ -1,19 +1,0 @@
-# Copyright (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .dataset_item import DatasetItem, DatasetItemAnnotation, DatasetItemFormat, DatasetItemSubset
-from .label import LabelReference
-from .shape import FullImage, Point, Polygon, Rectangle, Shape
-
-__all__ = [
-    "DatasetItem",
-    "DatasetItemAnnotation",
-    "DatasetItemFormat",
-    "DatasetItemSubset",
-    "FullImage",
-    "LabelReference",
-    "Point",
-    "Polygon",
-    "Rectangle",
-    "Shape",
-]
