+++ application/backend/app/schemas/dataset_item.py
@@ -6,9 +6,13 @@ from uuid import UUID
 
 from pydantic import BaseModel
 
-from app.core.models import BaseRequiredIDNameModel, Pagination
-from app.entities import DatasetItemAnnotation, DatasetItemFormat
-from app.entities.dataset_item import DatasetItemSubset
+from app.core.models import (
+    BaseRequiredIDNameModel,
+    DatasetItemAnnotation,
+    DatasetItemFormat,
+    DatasetItemSubset,
+    Pagination,
+)
 
 
 class DatasetItemView(BaseRequiredIDNameModel):
