+++ application/backend/tests/unit/services/test_dataset_service.py
@@ -4,8 +4,8 @@ from uuid import uuid4
 
 import pytest
 
+from app.core.models import DatasetItemAnnotation, FullImage, LabelReference, Point, Polygon, Rectangle
 from app.db.schema import DatasetItemDB
-from app.entities import DatasetItemAnnotation, FullImage, LabelReference, Point, Polygon, Rectangle
 from app.schemas.label import LabelView
 from app.schemas.project import ProjectView, TaskType, TaskView
 from app.services import DatasetService
