+++ application/backend/tests/unit/services/test_datumaro_converter.py
@@ -6,7 +6,7 @@ from uuid import UUID, uuid4
 import numpy as np
 import pytest
 
-from app.entities import (
+from app.core.models import (
     DatasetItem,
     DatasetItemAnnotation,
     DatasetItemFormat,
